All right.  So after a ridiculous amount of time spent debugging, I have determined a theory about what the error is.
When the Destroy(some gameobject) is called, the object is destroyed THE NEXT FRAME.  When multiple objects are picked up at one time, however, this messes everything up.  All I did to fix it was use a for loop.  I tried using DestroyImmediate, but it ***ing happened again!!!! Piece of ****.  Damn I'm tired.