Refactored a couple classes, and added a bit of new functionality to the CreateLevelItems class.  Items are now classified as required or optional.  If an object is required, it is instantiated ONCE, at one point in the level.  If an object is optional, it can be instantiated multiple times or not at all.
Also changed around NPC physics.  NPCs now have a finite mass, which makes it so they are no longer death traps for the player.  They also have a moderate friction collider attached, so they no longer slide around.  This is mainly due to the fact that CharacterBaseActionClass now has a coroutine attached that keeps a character at a constant velocity.