Added the new terrain images, and the colliders have been modified appropriately.  Also changed the particle effect a bit.
Couple new things need to happen in later commits.  First off, the mammoths freaking suck when exposed to actual physics.  They cannot be on top of the NPCs or players.  Second, new terrain needs to be added.  The level is way too short.  And third, the NPCs are the key parts to the level: they need to be instantiated at some random point.