Misc changes, the trees are now looking at home in the game (after they were resized).  For future reference, use birme.net for future batch image resizing: it was simple and easy to use.
New game idea:
You try to take the "shiny" from Thuk, but he refuses to give it to you.

Additionally, everyone in the game could live in a hut: a different scene.  But I wonder whether I could keep the game loaded in the background while in the hut, to avoid loading the thing all over again.