Added the snowball, but I think that I will not end up using it.  Also changed all of the soldiers on the boss segment to instantiators, so that they are updated along with the changes that have been made to previous soldiers.
Since the Platformer Art Deluxe Pack has igloo images, why not use the igloo as Thuk's home?  Make a couple for the people who sell stuff, and others for soldiers that lock the door and fight you.  Thuk's can be colored differently to emphasize that it is not the same.

Terrain Concept: Player kind of has to go a roundabout way (on purpose).

--------------          -----  /
               \              /
                 \           /
                   \--------/